requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
jinja2==3.1.3

# Face recognition (CPU-only)
insightface==0.7.3
onnxruntime==1.17.0
numpy==1.26.4

# Image processing
Pillow==10.2.0
rawpy==0.19.1
imageio==2.34.0

# Database
aiosqlite==0.19.0

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
aiofiles==23.2.1

